requirements.txt
Django==6.0.0
djangorestframework==3.16.1
psycopg[binary]
python-decouple==3.8
django-cors-headers==4.3.1
drf-spectacular==0.27.2
gunicorn==21.2.0

